新增支付失败界面，并完成重新付款逻辑。

+++ app/src/main/java/com/example/shanyu/main/home/bean/WxPayBean.java
@@ -20,7 +20,7 @@ public class WxPayBean implements Serializable {
     private String appid;
     private String partnerid;
     private Integer order_id;
-    private Integer money;
+    private double money;
     private String prepayid;
     private String noncestr;
     private String timestamp;
@@ -51,7 +51,7 @@ public class WxPayBean implements Serializable {
         this.order_id = orderId;
     }
 
-    public Integer getMoney() {
+    public double getMoney() {
         return money;
     }
